cleanup of old sector synth classes

+++ BizHawk.Emulation.DiscSystem/DiscUtils.cs
@@ -3,9 +3,9 @@ namespace BizHawk.Emulation.DiscSystem
 	public static class DiscUtils
 	{
 		/// <summary>
-		/// converts the given byte to a BCD value
+		/// converts the given int to a BCD value
 		/// </summary>
-		public static byte BCD_Byte(this byte val)
+		public static int BCD_Byte(this int val)
 		{
 			byte ret = (byte)(val % 10);
 			ret += (byte)(16 * (val / 10));
